Criei a classe GenreDatabase para lidar com a tabela de generos musicais
