+++ src/data/MusicDatabase.ts
@@ -1,8 +1,11 @@
 import { Genre, Music } from "../business/entities/Music";
 import { MySqlError } from "../business/error/MySqlError";
 import { BaseDatabase } from "./BaseDatabase";
+import { GenreDatabase } from "./GenreDatabase";
 
 export class MusicDatabase extends BaseDatabase {
+
+    private genreDatabase = new GenreDatabase();
 
     private static toMusicModel = (music: any, genres: Genre[]): Music => {
         return music && new Music(
@@ -55,34 +58,16 @@ export class MusicDatabase extends BaseDatabase {
         userId: string
     ): Promise<Music[]> => {
         try {
-            const musicResult = await BaseDatabase.connection(BaseDatabase.MUSIC_TABLE)
+            const musicResult = await BaseDatabase.connection()
                 .select('*')
+                .from(BaseDatabase.MUSIC_TABLE)
                 .where({ user_id: userId });
 
-            const musics: Music[] = [];  
+            const musics: Music[] = []; 
 
             for (let music of musicResult) {
-                const genreResult = await BaseDatabase.connection.raw(`
-                    SELECT genre_id as id, name
-                    FROM ${BaseDatabase.MUSIC_TABLE}
-                    JOIN ${BaseDatabase.MUSIC_GENRE_TABLE}
-                    ON ${BaseDatabase.MUSIC_GENRE_TABLE}.music_id = ${BaseDatabase.MUSIC_TABLE}.id
-                    JOIN ${BaseDatabase.GENRE_TABLE}
-                    ON ${BaseDatabase.GENRE_TABLE}.id = ${BaseDatabase.MUSIC_GENRE_TABLE}.genre_id
-                    WHERE ${BaseDatabase.MUSIC_TABLE}.id = '${music.id}'
-                `);
-
-                const genres: Genre[] = [];
-                
-                for (let genre of genreResult[0]) {
-                    genres.push({ 
-                        id: genre.id,
-                        name: genre.name 
-                    });
-                };
-
-                musics.push(MusicDatabase.toMusicModel(musicResult[0], genres));
-
+                const genres = await this.genreDatabase.selectGenreByMusic(music.id);
+                musics.push(MusicDatabase.toMusicModel(music, genres));
             };  
 
             return musics;
@@ -99,24 +84,7 @@ export class MusicDatabase extends BaseDatabase {
                 .select('*')
                 .where({ id });
 
-            const genres: Genre[] = [];    
-            
-            const genreResult = await BaseDatabase.connection.raw(`
-                SELECT genre_id as id, name
-                FROM ${BaseDatabase.MUSIC_TABLE}
-                JOIN ${BaseDatabase.MUSIC_GENRE_TABLE}
-                ON ${BaseDatabase.MUSIC_GENRE_TABLE}.music_id = ${BaseDatabase.MUSIC_TABLE}.id
-                JOIN ${BaseDatabase.GENRE_TABLE}
-                ON ${BaseDatabase.GENRE_TABLE}.id = ${BaseDatabase.MUSIC_GENRE_TABLE}.genre_id
-                WHERE ${BaseDatabase.MUSIC_TABLE}.id = '${id}'
-            `);
-
-            for (let genre of genreResult[0]) {
-                genres.push({ 
-                    id: genre.id,
-                    name: genre.name 
-                });
-            };    
+            const genres = await this.genreDatabase.selectGenreByMusic(id);    
 
             return MusicDatabase.toMusicModel(musicResult[0], genres);
